+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 **Next**
 
+* Nothing yet.
+
+**v0.16.0**
+
 * Add settings to control whether or not tooltips are shown on the 
   settings and help icons.
-  
+
 * Change all tippy-controlled content to use `light-border` theme.
 
 * Show tooltip when hovering over settings / help icons.